Dimensions now save and load things from disk
Also various bug fixes to the NBT code because I'm a moron and Copy
Pasting code is never a good idea ever while this universe exists

+++ utils/src/main/java/com/hackhalo2/nbt/tags/TagDouble.java
@@ -17,7 +17,7 @@ public final class TagDouble extends AbstractTag {
 	public TagDouble(NBTInputStream in, boolean anonymous) throws NBTException, IOException {
 		 super(in, anonymous, TagType.DOUBLE.getID());
 		 
-		 this.value = in.readShort();
+		 this.value = in.readDouble();
 	}
 	
 	public double getValue() {
